[core][telemetry] Drop vLLM-specific gauge override; fold unit tests
Gauges aggregate by sum generically. Ray core no longer names a metric
from another project. A ratio gauge overcounts under sum and should be
aggregated at the query layer until the metric owner can declare its
aggregation.

Move the unit tests into test_metric_cardinality.py so they run under the
existing Bazel target instead of an untargeted new file.
