+++ python/ray/tests/test_metric_cardinality.py
@@ -13,12 +13,14 @@ from ray._common.test_utils import (
     PrometheusTimeseries,
 )
 from ray._common.network_utils import build_address
+import ray._private.telemetry.metric_cardinality as mc
 from ray._private.telemetry.metric_cardinality import (
     WORKER_ID_TAG_KEY,
     REPLICA_ID_TAG_KEY,
     TASK_OR_ACTOR_NAME_TAG_KEY,
     MetricCardinality,
 )
+from ray._private.telemetry.metric_types import MetricType
 
 try:
     import prometheus_client
@@ -224,6 +226,81 @@ def test_serve_metric_cardinality(_setup_serve_metric_cluster, cardinality_level
             assert sample.value == 6.0, f"Expected summed 6.0, got {sample.value}"
 
     wait_for_assertion(_validate, timeout=30, retry_interval_ms=1000)
+
+
+# A Serve metric carries ReplicaId; a plain user metric does not.
+_SERVE_TAGS = ["model_name", REPLICA_ID_TAG_KEY, WORKER_ID_TAG_KEY]
+_USER_TAGS = ["model_name", WORKER_ID_TAG_KEY]
+
+
+@pytest.fixture
+def set_level():
+    """Set the cardinality level and clear the per-name label cache."""
+
+    def _set(level: str):
+        mc._CARDINALITY_LEVEL = MetricCardinality(level)
+        mc._HIGH_CARDINALITY_LABELS.clear()
+
+    original = mc._CARDINALITY_LEVEL
+    yield _set
+    mc._CARDINALITY_LEVEL = original
+    mc._HIGH_CARDINALITY_LABELS.clear()
+
+
+def _drop(name, tag_keys):
+    return MetricCardinality.get_high_cardinality_labels_to_drop(name, tag_keys)
+
+
+def test_unit_legacy_drops_nothing(set_level):
+    set_level("legacy")
+    assert _drop("tasks", _USER_TAGS) == []
+    assert _drop("vllm_foo", _SERVE_TAGS) == []
+
+
+@pytest.mark.parametrize("level", ["recommended", "low"])
+def test_unit_serve_metric_drops_worker_and_replica_id(set_level, level):
+    set_level(level)
+    assert _drop("vllm_foo", _SERVE_TAGS) == [WORKER_ID_TAG_KEY, REPLICA_ID_TAG_KEY]
+
+
+def test_unit_plain_user_metric_is_untouched(set_level):
+    # No ReplicaId and not a Ray high-cardinality metric -> nothing dropped.
+    for level in ("recommended", "low"):
+        set_level(level)
+        assert _drop("my_app_metric", _USER_TAGS) == []
+
+
+def test_unit_tasks_actors_follow_existing_rules(set_level):
+    set_level("recommended")
+    assert _drop("tasks", _USER_TAGS) == [WORKER_ID_TAG_KEY]
+    set_level("low")
+    assert _drop("actors", _USER_TAGS) == [
+        WORKER_ID_TAG_KEY,
+        TASK_OR_ACTOR_NAME_TAG_KEY,
+    ]
+
+
+def test_unit_name_only_call_is_not_cached(set_level):
+    # A call without tag_keys must not poison the cache for a Serve metric.
+    set_level("recommended")
+    assert _drop("vllm_foo", None) == []
+    assert _drop("vllm_foo", _SERVE_TAGS) == [WORKER_ID_TAG_KEY, REPLICA_ID_TAG_KEY]
+
+
+def test_unit_gauge_defaults_to_sum(set_level):
+    agg = MetricCardinality.get_aggregation_function("vllm_foo", MetricType.GAUGE)
+    assert agg([3.0, 3.0]) == 6.0
+
+
+def test_unit_counter_and_sum_always_sum(set_level):
+    for metric_type in (MetricType.COUNTER, MetricType.SUM):
+        agg = MetricCardinality.get_aggregation_function("vllm_foo", metric_type)
+        assert agg([1.0, 2.0, 3.0]) == 6.0
+
+
+def test_unit_histogram_has_no_aggregation_function(set_level):
+    with pytest.raises(ValueError):
+        MetricCardinality.get_aggregation_function("vllm_foo", MetricType.HISTOGRAM)
 
 
 if __name__ == "__main__":
